✨ feat - Refreshes the new wallet selection page.
