+++ Source/Mobile/Finance/Finance.csproj.user
@@ -38,7 +38,7 @@
     <MauiXaml Update="Pages\DangerZone\Delete\DeleteWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Pages\Walleting\CreateWallet\CreateWalletPage.xaml">
+    <MauiXaml Update="Pages\Walleting\CreateWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
     <MauiXaml Update="Pages\Historic\HistoricPage.xaml">
@@ -53,7 +53,7 @@
     <MauiXaml Update="Pages\LoadingPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Pages\SelectWallet\SelectWalletPage.xaml">
+    <MauiXaml Update="Pages\Walleting\SelectWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
     <MauiXaml Update="Resources\Styles\MultilineEditStyle.xaml">
