+++ Source/Mobile/Finance/Pages/Historic/HistoricPage.xaml.cs
@@ -11,7 +11,7 @@ public partial class HistoricPage : ContentPage {
     protected override async void OnAppearing() {
         if(!viewModel.IsRunning) {
             viewModel.IsRunning = true;
-            await Task.Delay(500);
+            await Task.Delay(100);
             viewModel.IsRunning = false;
         }
     }
