+++ Source/Mobile/Finance/Pages/Strategy/StrategyViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.Input;
 
 namespace Finance.Pages.Strategy;
